+++ lib/screens/home_screen/home_coach_mark.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:krosscutting_app/screens/home_page/home_button.dart';
+import 'package:krosscutting_app/screens/home_screen/home_button.dart';
 import 'package:krosscutting_app/widgets/moving_icon.dart';
 
 class HomeCoachMark extends StatefulWidget {
